Fixes a few issues
The bullet error was because you misspelled position
The bullet was colliding with the ship because the nose was too short
There were several problems that prevented the enemy from shooting (the Timer was missing a signal, the "nose" was much too short, the enemy was shooting in the wrong direction)
The explosion animation was set to loop; I unchecked that. I also started it on frame 0 instead of frame 4.
Let me know if I can help with anything else!